more code cleanups

+++ src/UserInterface/NSButton+Color.m
@@ -26,7 +26,7 @@
 - (void)setTextColor:(NSColor *)textColor {
     NSMutableAttributedString *attrTitle = [[NSMutableAttributedString alloc] 
                                             initWithAttributedString:[self attributedTitle]];
-    int len = [attrTitle length];
+    NSUInteger len = [attrTitle length];
     NSRange range = NSMakeRange(0, len);
     [attrTitle addAttribute:NSForegroundColorAttributeName 
                       value:textColor
